test passing

+++ dynts/utils/section.py
@@ -17,16 +17,17 @@ def scalarasiter(x):
     else:
         return (x,)
 
+
 def asarray(x, dtype=None):
     '''Convert ``x`` into a ``numpy.ndarray``.'''
     iterable = scalarasiter(x)
     if isinstance(iterable, ndarray):
         return iterable
     else:
-        if not hasattr(iterable,'__len__'):
+        if not hasattr(iterable, '__len__'):
             iterable = list(iterable)
         if dtype == object_type:
-            a = ndarray((len(iterable),), dtype = dtype)
+            a = ndarray((len(iterable),), dtype=dtype)
             for i,v in enumerate(iterable):
                 a[i] = v
             return a
